+++ packages/tiptap-extensions/web/src/paragraph.ts
@@ -3,7 +3,7 @@ import { mergeAttributes } from '@tiptap/core';
 import type { Node as ProsemirrorNode } from 'prosemirror-model';
 import paragraphDataDefaults from 'ricos-schema/dist/statics/paragraph.defaults.json';
 import type { RicosExtension, DOMOutputSpec } from 'ricos-tiptap-types';
-import { Node_Type } from 'ricos-schema';
+import { Node_Type, TextStyle_TextAlignment } from 'ricos-schema';
 
 export interface ParagraphOptions {
   HTMLAttributes: Record<string, unknown>;
@@ -57,10 +57,19 @@ export const paragraph: RicosExtension = {
 
       renderHTML({ HTMLAttributes, node }) {
         const styles = createStyleAttribute(node);
+        const attrs = mergeAttributes(this.options.HTMLAttributes, HTMLAttributes, styles);
+        const textAlignment = attrs?.textStyle?.textAlignment;
+        const shouldAddDir =
+          textAlignment === TextStyle_TextAlignment.LEFT ||
+          textAlignment === TextStyle_TextAlignment.RIGHT;
         return [
           'div',
-          mergeAttributes(this.options.HTMLAttributes, HTMLAttributes, styles),
-          ['span', { style: 'display: inline-block;' }, 0],
+          attrs,
+          [
+            'span',
+            { style: 'display: inline-block;', ...(shouldAddDir ? { dir: 'auto' } : {}) },
+            0,
+          ],
         ] as DOMOutputSpec;
       },